cardno change  from double to long

+++ Code/FraudDetection/fraudanalytics-dto/src/main/java/com/masteklabs/fraudanalytics/common/dto/PredictionResult.java
@@ -18,7 +18,7 @@ public class PredictionResult implements Serializable{
 	private Long accountId;
 	private Long transactionTime;
 	private Double amount;
-	private Double cardNumber; 
+	private Long cardNumber; 
 
 
 	public Long getAccountId() {
